+++ app/controllers/videos_controller.rb
@@ -14,23 +14,4 @@ class VideosController < ApplicationController
     @videos = Video.search_by_title(params[:search])
   end
 
-  def review
-    @video = Video.find(params[:id])
-    @review = Review.new(review_params)
-    @review.user = current_user
-    @review.video = @video
-
-    if @review.save
-      flash[:success] = "Your review has been recorded"
-      redirect_to video_path(@video)
-    else
-      render :show
-    end
-  end
-
-  private
-  def review_params
-    params.require(:review).permit(:rating, :review)
-  end
-
 end
